+++ Concordia_Campus_App/.gitignore
@@ -36,3 +36,6 @@ yarn-error.*
 *.tsbuildinfo
 
 .env
+
+package-lock.json
+package.json
